reorganized files, cleanup player movement
