+++ backend/pong/src/pong.html
@@ -47,7 +47,7 @@
 		p2CornerCol =	"red"
 		
 		/*game vars*/
-		singlePlayer =	"true"
+		singlePlayer =	"false"
 		ballSpeed =		"standard"	/*glacial, slow, standard, fast, insane*/
 		ballSize = 		"normal"	/*tiny, small, normal, big, huge*/
 		paddleSpeed =	"standard"	/*glacial, slow, standard, fast, insane*/
